refactor: Rename rules to ruleUtils to avoid name clash with Rules
The problem was visible during tests since macOS filesystem is case
sensitive.
