+++ src/ducks/settings/ruleUtils.js
